Essential steps for supporting Windows and more
Modified parts of the code that gcc/clang supported but not MSVC, to
make MSVC support them too; fixed some bugs; made entry header
encoding/decoding more platform-independent; modified entry header unit
test code too for that; setup installation procedure for installing
squeeze on windows; few more changes.

+++ src/misc/cpu_info.cpp
@@ -7,8 +7,8 @@ namespace squeeze::misc {
 
 std::size_t get_nr_available_cpu_cores()
 {
-    int process_affinity_mask;
-    int system_affinity_mask;
+    DWORD64 process_affinity_mask;
+    DWORD64 system_affinity_mask;
 
     if (!GetProcessAffinityMask(GetCurrentProcess(), &process_affinity_mask, &system_affinity_mask))
         return 1;
